backend/requirements.txt
alembic==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
asttokens==2.4.1
asyncpg==0.29.0
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
brotli==1.2.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==46.0.6
cssselect2==0.9.0
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
executing==2.1.0
fastapi==0.115.0
Flask==3.1.0
Flask-SocketIO==5.5.1
fonttools==4.62.1
fpdf2==2.8.7
git-filter-repo==2.47.0
greenlet==3.1.1
groq==0.11.0
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
icecream==2.1.3
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.54.0
packaging==26.0
passlib==1.7.4
pillow==12.2.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pyasn1==0.6.3
pycparser==3.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pydyf==0.12.1
Pygments==2.18.0
PyJWT==2.9.0
pyotp==2.9.0
pyphen==0.17.2
pytest==9.0.2
python-dotenv==1.0.0
python-engineio==4.11.2
python-jose==3.3.0
python-multipart==0.0.9
python-socketio==5.12.1
pytz==2026.1.post1
PyYAML==6.0.3
qrcode==8.2
redis==5.1.0
requests==2.32.3
resend==2.29.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.38.6
tinycss2==1.5.1
tinyhtml5==2.1.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.30.0
watchfiles==1.1.1
weasyprint==68.1
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.3
wsproto==1.2.0
zopfli==0.4.1
